+++ server.js
@@ -13,14 +13,14 @@ const clubHeadRouter = require('./routes/club_head')
 const clubMembersRouter = require('./routes/club_members')
 const clubsRouter = require('./routes/club')
 const eventsRouter = require('./routes/events')
-const imagesRouter = require('./routes/images')
+const imagesRouter = require('./routes/supporting/images')
 const fileRouter = require('./routes/files')
 const projectRouter = require('./routes/projects')
 const achievementsRouter = require('./routes/achievements')
-const notifyRouter = require('./routes/notify')
+const notifyRouter = require('./routes/supporting/notify')
 const blogRouter = require('./routes/blog')
 const newsRouter = require('./routes/news')
-const registerRouter = require('./routes/register')
+const registerRouter = require('./routes/supporting/register')
 
 const app = express()
 const port = process.env.PORT || 5000
